Change from "GPropItemParam* param" into "GPropItemParam* param"

+++ src/net/capture/gnetfilter.h
@@ -20,9 +20,9 @@
 struct GNetFilter : GCapture {
   Q_OBJECT
   Q_PROPERTY(int queueNum MEMBER queueNum_)
-  Q_PROPERTY(size_t snapLen MEMBER snapLen_)
+  Q_PROPERTY(int snapLen MEMBER snapLen_)
   Q_PROPERTY(Verdict acceptVerdict MEMBER acceptVerdict_)
-  Q_PROPERTY(uint32_t mark MEMBER mark_)
+  Q_PROPERTY(quint32 mark MEMBER mark_)
   Q_ENUMS(Verdict)
 
 public:
